+++ example_test.go
@@ -58,3 +58,34 @@ func ExampleAdaptOne() {
 	// Output:
 	// [1 3 5]
 }
+
+func ExampleReturn() {
+	// Return binds return value to a function without parameters.
+	// It can be used for passing static value to functions that accept
+	// functions.
+	f := func(fn func() string) {
+		fmt.Println(fn())
+	}
+	f(curry.Return("message"))
+	// Output: message
+}
+
+func ExampleLazyOne0() {
+
+	process := func(expensive string) {
+		fmt.Println("That was", expensive)
+	}
+	expensive := func() string {
+		fmt.Println("Computing...")
+		return "expensive "
+	}
+
+	defer curry.LazyOne0(process)(expensive)
+
+	fmt.Println("The expensive process argument hasn't been computed yet.")
+
+	// Output:
+	// The expensive process argument hasn't been computed yet.
+	// Computing...
+	// That was expensive
+}