+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 
 - Adapt* functions.
+- Lazy* and Thunk* functions.
 
 ## [0.0.1] - Initial release